+++ packages/ricos-editor/web/src/toolbars/FloatingAddPluginMenu.tsx
@@ -1,17 +1,19 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { CirclePlusIcon as PlusIcon } from 'wix-rich-content-ui-components';
 import styles from '../../statics/styles/floating-add-plugin-menu.scss';
 import { posToDOMRect } from './utils/posToDOMRect';
+import classNames from 'classnames';
 
-const FloatingAddPluginMenu = props => {
-  const [position, setPosition] = useState({ top: '0px', left: '-40px' });
-  const { editor } = props;
+const FloatingAddPluginMenu = ({ editor, isMobile, languageDir }) => {
+  const [position, setPosition] = useState({});
+  const floatingMenuWrapperRef = useRef<HTMLDivElement>(null);
 
   const handleButtonPosition = () => {
     const { selection } = editor.state;
     const { top } = posToDOMRect(editor.view, selection.from, selection.to);
-    const topPos = top < 53 ? position.top : `${top - 100}px`;
-    setPosition({ ...position, top: topPos });
+    const offsetTop = floatingMenuWrapperRef?.current?.getBoundingClientRect().top as number;
+    const verticalPos = { top: `${top - offsetTop}px` };
+    setPosition({ ...verticalPos });
   };
 
   useEffect(() => {
@@ -19,8 +21,19 @@ const FloatingAddPluginMenu = props => {
   }, [editor.state]);
 
   return (
-    <div className={styles.floatingAddPluginMenu_wrapper}>
-      <PlusIcon className={styles.floatingAddPluginMenu_button} style={{ ...position }} />
+    <div
+      dir={languageDir}
+      className={styles.floatingAddPluginMenu_wrapper}
+      ref={floatingMenuWrapperRef}
+    >
+      <button
+        className={classNames(styles.floatingAddPluginMenu_button, {
+          [styles.floatingAddPluginMenu_button_mobile]: isMobile,
+        })}
+        style={position}
+      >
+        <PlusIcon />
+      </button>
     </div>
   );
 };